Split transactions HOC
